+++ getEC2.js
