+++ package.json
@@ -4,7 +4,7 @@
   "description": "Provides Promises/A+ compliant version of all your favorite AWS SDK clients.",
   "main": "index.js",
   "scripts": {
-    "test": "node_modules/.bin/tape test/**/*.js"
+    "test": "node_modules/.bin/tape test/**.js"
   },
   "repository": {
     "type": "git",
